refactor(schema): refactor to use owner.schema everywhere
