+++ schemas/owners.schema.js
@@ -62,4 +62,4 @@ const ownerSchema = new mongoose.Schema({
 
 const Owner = mongoose.model('Owner', ownerSchema);
 
-export default Owner;
+module.exports = Owner;
